backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
passlib==1.7.4
python-multipart==0.0.6
openai>=1.12.0
httpx>=0.24.1
aiohttp>=3.8.0 